Reorg asset service
